+++ include/RMF/internal/utility.h
@@ -10,20 +10,23 @@
 
 #include <RMF/config.h>
 
-RMF_ENABLE_WARNINGS namespace RMF {
-  class AliasFactory;
-  class NodeHandle;
-  class NodeConstHandle;
+RMF_ENABLE_WARNINGS
+namespace RMF {
+namespace decorator {
+class AliasFactory;
+}
+class NodeHandle;
+class NodeConstHandle;
 
-  namespace internal {
-  template <class VA, class VB>
-  void copy_vector(const VA& va, VB& vb) {
-    vb.insert(vb.end(), va.begin(), va.end());
-  }
+namespace internal {
+template <class VA, class VB>
+void copy_vector(const VA& va, VB& vb) {
+  vb.insert(vb.end(), va.begin(), va.end());
+}
 
-  RMFEXPORT NodeHandle add_child_alias(AliasFactory af, NodeHandle parent,
-                                       NodeConstHandle child);
-  }
+RMFEXPORT NodeHandle add_child_alias(decorator::AliasFactory af,
+                                     NodeHandle parent, NodeConstHandle child);
+}
 }
 
 RMF_DISABLE_WARNINGS